+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/decode/DecodeConfig.common.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2024 panpf <panpfpanpf@outlook.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.github.panpf.sketch.decode
+
+import com.github.panpf.sketch.request.ImageRequest
+
+/**
+ * Decode configuration
+ *
+ * @see com.github.panpf.sketch.core.android.test.decode.DecodeConfigTest
+ * @see com.github.panpf.sketch.core.nonandroid.test.decode.DecodeConfigTest
+ */
+expect class DecodeConfig
+
+/**
+ * Create a DecodeConfig based on the parameters related to image quality in the request
+ *
+ * @see com.github.panpf.sketch.core.android.test.decode.DecodeConfigTest.testDecodeConfig
+ * @see com.github.panpf.sketch.core.nonandroid.test.decode.DecodeConfigTest.testDecodeConfig
+ */
+expect fun DecodeConfig(
+    request: ImageRequest,
+    mimeType: String? = null,
+    isOpaque: Boolean = false
+): DecodeConfig