+++ sketch-core/src/nonAndroidMain/kotlin/com/github/panpf/sketch/decode/DecodeConfig.kt
@@ -1,3 +1,19 @@
+/*
+ * Copyright (C) 2024 panpf <panpfpanpf@outlook.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.github.panpf.sketch.decode
 
 import com.github.panpf.sketch.request.ImageRequest
@@ -9,10 +25,10 @@ import org.jetbrains.skia.ColorType
  *
  * @see com.github.panpf.sketch.core.nonandroid.test.decode.DecodeConfigTest.testDecodeConfig
  */
-fun DecodeConfig(
+actual fun DecodeConfig(
     request: ImageRequest,
-    mimeType: String? = null,
-    isOpaque: Boolean = false,
+    mimeType: String?,
+    isOpaque: Boolean,
 ): DecodeConfig = DecodeConfig().apply {
     val userColorType = request.colorType?.getColorType(mimeType, isOpaque)?.colorType
     if (userColorType != null) {
@@ -30,7 +46,7 @@ fun DecodeConfig(
  *
  * @see com.github.panpf.sketch.core.nonandroid.test.decode.DecodeConfigTest
  */
-data class DecodeConfig(
+actual data class DecodeConfig(
     /**
      * If set to a value > 1, requests the decoder to subsample the original
      * image, returning a smaller image to save memory. The sample size is